requirements.txt
flask
pandas
scrapy
scrapy_splash
flask_mysqldb
python-dotenv